# Conflicts:
#	src/test/java/org/jacop/SmallDenseDomainTest.java

+++ src/main/java/org/jacop/core/IntVar.java
@@ -1,4 +1,4 @@
-/**
+/*
  * IntVar.java
  * This file is part of JaCoP.
  * <p>
@@ -40,7 +40,7 @@ import org.jacop.satwrapper.translation.SatCPBridge;
  * Defines a Finite Domain Variable (FDV) and related operations on it.
  *
  * @author Radoslaw Szymanek and Krzysztof Kuchcinski
- * @version 4.4
+ * @version 4.5
  */
 
 public class IntVar extends Var {
@@ -144,7 +144,7 @@ public class IntVar extends Var {
      */
     public IntVar(Store store, String name, int min, int max) {
 
-        if (max - min > 63)
+        if ((long)max - (long)min > 63L)
             commonInitialization(store, name, new IntervalDomain(min, max));
         else
             commonInitialization(store, name, new SmallDenseDomain(min, max));
